fix: Clean codes #189
