+++ frontend/src/routes/RouterPage.tsx
@@ -61,7 +61,6 @@ import Docs from "pages/docs/Index";
 import ChooseAcademy from "pages/login/ChooseAcademy";
 import Dev from "pages/dev/Index";
 import Schedule from "pages/dev/Schedule";
-import Workspace from "pages/dev/Workspace/Index";
 
 function RouterPage() {
   const { currentUser } = useAuth();
@@ -440,7 +439,6 @@ function RouterPage() {
               <Route path="test" element={<Test />}></Route>
               <Route path="e" element={<E />}></Route>
               <Route path="schedule" element={<Schedule />}></Route>
-              <Route path="workspace" element={<Workspace />}></Route>
               {/*  404 error */}
               <Route path="*" element={<Http404 />}></Route>
             </Route>
