+++ frontend/src/types/auth.ts
@@ -131,12 +131,3 @@ export type TSeason = {
   //____________________//
   registrations: TSeasonRegistration[];
 };
-
-export type TWorkspace = {
-  id: string;
-  email: string;
-  accessToken: string;
-  expires: Date;
-  refreshToken: string;
-  calendars?: { items: any[] };
-};
